requirements.txt
labelme
ultralytics
json
shutil
scikit-learn
numpy
opencv-python
albumentations
Flask
pillow 
torch 
torchvision